pyproject.toml
[project]
name = "seekerbot"
version = "0.1.0"
description = "JobMiner - Rastreador de Vagas"
authors = [
    { name = "Caio Mussatto", email = "Caio.mussatto@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "markdown>=3.10",
    "openai>=2.24.0",
    "psycopg2-binary>=2.9.11",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.1",
    "python-jobspy>=1.1.82",
    "sqlalchemy>=2.0.45",
]